Add utxo caching everywhere and define BLOCK_CONFIRMATIONS for smartfee reqs
